Print body as a concatenation of the common labels.
With the existing code we're having the issue that it ends up creating
empty message bodies. These are rejected by MessageBird.

With this change we simply look at all of the labels that are common
across the entire message. As AlertManager generates a single request
for every different alert name, it means in practice that you get
something along the lines of:

  TooManyGophers | localhost:9091 | pushprom | codelab-monitor |
  gophers is bigger than 10

By sending a single SMS, we also cut down on costs.